add to cart && add to favourite

+++ app/src/main/java/com/example/shopapp/models/CartItemData.kt
@@ -3,8 +3,8 @@ package com.example.shopapp.models
 data class CartItemData(
     val cart_items: List<CartItem>?=null,
     val cartItem:CartItem?=null,
-    val sub_total: Int,
-    val total: Int
+    val sub_total: Double,
+    val total: Double
 )
 
 //AuthResponse<CartItemData>
